backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0

# Document parsing
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# Vector store
chromadb==0.4.18
faiss-cpu==1.9.0
numpy==1.26.4

# HTML parsing
beautifulsoup4==4.12.2

# HTML to PDF conversion
weasyprint==60.2
pydyf<0.11.0  # Pin compatible version (0.11.0 has breaking changes)

# Tokenization (optional but recommended)
tiktoken==0.5.1

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
httpx==0.25.2

# NeMo Guardrails
nemoguardrails>=0.5.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Database
sqlalchemy>=2.0.0

# Web scraping and async HTTP
aiohttp>=3.9.0
feedparser>=6.0.10
